requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
anthropic==0.25.0
httpx==0.27.0
PyJWT==2.8.0
python-dotenv==1.0.1
